Run black on test_hillshade.py
